+++ lib/churn_vs_complexity.rb
@@ -12,6 +12,7 @@ require_relative 'churn_vs_complexity/churn'
 require_relative 'churn_vs_complexity/cli'
 require_relative 'churn_vs_complexity/config'
 require_relative 'churn_vs_complexity/serializer'
+require_relative 'churn_vs_complexity/git_date'
 
 module ChurnVsComplexity
   class Error < StandardError; end