+++ test/churn_vs_complexity/serializer/csv_test.rb
@@ -10,7 +10,7 @@ module ChurnVsComplexity
           'file1' => [1, 1],
           'file2' => [2, 16],
         }
-        result = CSV.serialize(values_by_file)
+        result = CSV.serialize({ values_by_file: })
         assert_equal "file1,1,1\nfile2,2,16\n", result
       end
     end