+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## [1.0.0] - 2024-06-07
 
 - Initial release
+
+## [1.1.0] - 2024-09-20
+
+- Introduce `--summary` flag to output summary statistics for churn and complexity
+- Introduce `--month`, `--quarter`, and `--year` short-hand flags to calculate churn for different time periods relative to the most recent commit